Update HWTC APIs and improve parsing logic for channel data
This update adjusts the handling of HWTC APIs with the ability to retrieve user tokens and apply them in the channel data request. Also revised the parsing logic for input in tests and regular expressions in `bytesToChannels` for both hwtc and zteg. An EPGPath field is added to the config and used in `getChannelBytes` and `getEPGBytes`. An `epgPath` and `userToken` fields are added to the `Client` struct for better organization of data. Additionally, the `goquery` library was added for HTML parsing.