Update Docker tags and ldflags in GitHub workflows
This commit updates the Docker image tags in the docker-dev.yaml and docker-release.yaml GitHub workflows. It sets the specific repository name instead of using the variable for the current actor. In addition, it modifies the ldflags settings in release.yaml to use the name of the GitHub actor. These changes aim to streamline the image creation and release processes.